updated EyeCommander to include new calibration subpackage
